Fixed crashes when adding and deleting books
The Book entity was missing a default date for the copyright attribute.
The nav bar button IBAction funcs for the add and cancel buttons were defined incorrectly.
The fetch controller changed delegate for a delete was using newIndexPath which is nil instead of the indexPath.

Thanks for coding this! I'm thinking of rewriting it in SwiftUI.